DDS2-2078: Allowing users to enter decimal places in the penalty percentage fields..

+++ app/models/offshore/TaxYearLiabilities.scala
@@ -24,7 +24,7 @@ final case class TaxYearLiabilities(
   capitalGains: BigInt,
   unpaidTax: BigInt,
   interest: BigInt,
-  penaltyRate: Int,
+  penaltyRate: BigDecimal,
   penaltyRateReason: String,
   foreignTaxCredit: Boolean
 )
